fix: resolve Scala.js type interop issues in npm package
  - Change SlotConfig to use Double instead of Long for JS compatibility
  - Use Uint8Array for byte arrays in evalPlutusScripts API
  - Add @nowarn for intentional Long→Double conversion in getInterval
  - Update TypeScript definitions with proper types
  - Bump sbt-scalajs to 1.20.2

+++ scalus-cardano-ledger/shared/src/main/scala/scalus/cardano/txbuilder/TxBuilder.scala
@@ -930,7 +930,7 @@ case class TxBuilder(
       *   the earliest valid time for the transaction
       */
     def validFrom(from: Instant): TxBuilder = {
-        val slot = env.slotConfig.timeToSlot(from.toEpochMilli)
+        val slot = env.slotConfig.timeToSlot(from.toEpochMilli).toLong
         addSteps(TransactionBuilderStep.ValidityStartSlot(slot))
     }
 
@@ -943,7 +943,7 @@ case class TxBuilder(
       *   the latest valid time for the transaction (exclusive)
       */
     def validTo(to: Instant): TxBuilder = {
-        val slot = env.slotConfig.timeToSlot(to.toEpochMilli)
+        val slot = env.slotConfig.timeToSlot(to.toEpochMilli).toLong
         addSteps(TransactionBuilderStep.ValidityEndSlot(slot))
     }
